Update Profitloss.java
This Java program calculates profit or loss based on the input of cost price (CP) and selling price (SP) entered by the user. If SP equals CP, it signifies no profit or loss. If SP is greater than CP, it denotes a profit, and if SP is less than CP, it signifies a loss.